Implement KerasModel `log_batches` with a separate callback for performance
Keras callbacks that run on every batch can slow down the training
of a model. So we will have separate internal callbacks
for batches and epochs.
