+++ scalarstop/model.py
@@ -144,6 +144,7 @@ import numpy as np
 import tensorflow as tf
 
 from scalarstop._filesystem import rmtree
+from scalarstop._keras_callbacks import BatchLoggingCallback, EpochCallback
 from scalarstop.datablob import DataBlob
 from scalarstop.exceptions import IsNotImplemented, ModelNotFoundError
 from scalarstop.model_template import ModelTemplate
@@ -330,100 +331,6 @@ class Model:
 _KERAS_HISTORY_TYPE = Dict[str, List[float]]
 
 
-def _logs_as_floats(logs: Mapping[str, Any]) -> Dict[str, float]:
-    """Convert Keras metric log values to floats."""
-    return {name: float(value) for name, value in logs.items()}
-
-
-class _ScalarStopKerasCallback(tf.keras.callbacks.Callback):
-    """A Keras callback to handle some of the bookkeeping."""
-
-    def __init__(
-        self,
-        *,
-        scalarstop_model: "KerasModel",
-        models_directory: Optional[str] = None,
-        train_store: Optional[TrainStore] = None,
-        log_batches: bool = False,
-        log_epochs: bool = False,
-        logger: Optional[Any] = None,
-    ):
-        super().__init__()
-        self._scalarstop_model = scalarstop_model
-        self._models_directory = models_directory
-        self._train_store = train_store
-        self._log_batches = log_batches
-        self._log_epochs = log_epochs
-        self._logger = logger or _LOGGER
-
-    def on_train_batch_end(  # pylint: disable=signature-differs
-        self, batch: int, logs: Dict[str, Any]
-    ) -> None:
-        """Enable issuing log messages at the end of every batch."""
-        super().on_train_batch_end(batch=batch, logs=logs)
-        # Log the batch.
-        if self._log_batches:
-            float_logs = _logs_as_floats(logs)
-            self._logger.info(
-                "Trained batch %s for epoch %s for model %s",
-                batch,
-                self._scalarstop_model.current_epoch,
-                self._scalarstop_model.name,
-                extra=dict(
-                    current_batch=batch,
-                    current_epoch=self._scalarstop_model.current_epoch,
-                    model_name=self._scalarstop_model.name,
-                    training_metrics=float_logs,
-                ),
-            )
-
-    def on_epoch_end(  # pylint: disable=signature-differs
-        self, epoch: int, logs: Dict[str, Any]
-    ) -> None:
-        """
-        Enable various tasks at the end of every epoch, such as:
-         - saving the model to the filesystem.
-         - saving epoch metrics to the TrainStore.
-         - logging epoch metrics to a Python logger.
-        """
-        super().on_epoch_end(epoch=epoch, logs=logs)
-        # Make sure that metrics are floats and not some
-        # unserializable data type like tf.Tensor
-        float_logs = _logs_as_floats(logs)
-
-        # Append epoch metrics to the model history.
-        for metric, value in float_logs.items():
-            if metric in self._scalarstop_model._history:
-                self._scalarstop_model._history[metric].append(value)
-            else:
-                self._scalarstop_model._history[metric] = [value]
-
-        # Save the model to the filesystem.
-        if self._models_directory:
-            self._scalarstop_model.save(self._models_directory)
-
-        if self._train_store:
-            self._train_store.insert_model_epoch(
-                epoch_num=self._scalarstop_model.current_epoch,
-                model_name=self._scalarstop_model.name,
-                metrics=float_logs,
-                ignore_existing=True,
-            )
-
-        # Log the epoch.
-        if self._log_epochs:
-            self._logger.info(
-                "Trained epoch %s for model %s",
-                self._scalarstop_model.current_epoch,
-                self._scalarstop_model.name,
-                extra=dict(
-                    current_epoch=self._scalarstop_model.current_epoch,
-                    model_name=self._scalarstop_model.name,
-                    training_metrics=float_logs,
-                ),
-            )
-
-
 class KerasModel(Model):
     """Trains :py:mod:`tf.keras` machine learning models generated by a :py:class:`~scalarstop.model_template.ModelTemplate` on the training and validation sets in a :py:class:`~scalarstop.datablob.DataBlob`."""  # pylint: disable=line-too-long
 
@@ -593,21 +500,35 @@ class KerasModel(Model):
             if verbose is None:
                 verbose = 1
 
+            # We start by adding any Keras callbacks the user wanted to add.
             if callbacks:
                 callbacks = list(callbacks)
             else:
                 callbacks = []
 
+            # Then we add our default Keras callback that handles all
+            # kinds of logging and saving tasks.
             callbacks.append(
-                _ScalarStopKerasCallback(
+                EpochCallback(
                     scalarstop_model=self,
+                    logger=logger or _LOGGER,
                     models_directory=models_directory,
-                    log_batches=log_batches,
                     log_epochs=log_epochs,
                     train_store=train_store,
-                    logger=logger,
                 )
             )
+
+            # Logging individual batches can significantly slow down
+            # a training process, so if the user wants to log batches,
+            # we use a separate callback. If the user does NOT want to
+            # log batches, we save a (costly) Python function call.
+            if log_batches:
+                callbacks.append(
+                    BatchLoggingCallback(
+                        scalarstop_model=self,
+                        logger=logger or _LOGGER,
+                    )
+                )
 
             if tensorboard_directory:
                 callbacks.append(
